Serialize HTML form elements into JS object
This package exports a function that inputs an HTML form element and
outputs a java script object adhering to the name, value, and type
properties of each element belonging to that form. The returned object
nests following an obsolete specification[1].

In addition to the old specs, we added the basic types of `number` and
`boolean` to the outputs.

* Inputs of type `checkbox` with no explicit value attribute get the
  value `true` if it is checked, and `false` otherwise.

* Inputs of type `number` and `range` gets it value casted to `number`.

* All other elements get their value as `string`.

[1]: https://www.w3.org/TR/html-json-forms/